fix for login and register errors.
