+++ src/CarService.UserAPI/Infrastructure/MyException.cs
@@ -1,0 +1,24 @@
+﻿using System.Runtime.Serialization;
+
+namespace CarService.UserAPI.Infrastructure
+{
+    [Serializable]
+    public class MyException : Exception
+    {
+        public MyException()
+        {
+        }
+        public MyException(string message)
+            : base(message)
+        {
+        }
+        public MyException(string message, Exception innerException)
+            : base(message, innerException)
+        {
+        }
+        protected MyException(SerializationInfo info, StreamingContext context)
+            : base(info, context)
+        {
+        }
+    }
+}
